Crewai AI Agents/ai_latest_development/pyproject.toml
[project]
name = "ai_latest_development"
version = "0.1.0"
description = "ai-latest-development using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai==0.11.0",
    "pyyaml>=6.0.1",
    "langchain==0.1.0",
    "langchain-core==0.1.0",
    "langchain-community==0.0.10",
    "python-dotenv>=1.0.0",
    "langchain-google-genai==0.0.6",
    "google-generativeai==0.3.2"
]

[project.scripts]
ai_latest_development = "ai_latest_development.main:run"
run_crew = "ai_latest_development.main:run"
train = "ai_latest_development.main:train"
replay = "ai_latest_development.main:replay"
test = "ai_latest_development.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
